+++ clients/derivatives-trading-options/src/websocket-streams/websocket-streams.ts
@@ -44,6 +44,7 @@ export class WebsocketStreams {
             ...(mode && { mode }),
             ...(poolSize && { poolSize }),
         });
+        websocketBase.streamIdIsStrictlyNumber = true;
         const websocketStreamsConnection = new WebsocketStreamsConnection(websocketBase);
         await websocketBase.connect(stream);
         return websocketStreamsConnection;